Emacs: fix keybindings, lexical-let instead let
I didn't know that elisp only has for default dynamic binding. So closures doesn't works as expected on this language. I always write my lisp thinking about Common Lisp, so is hard to understand something like that (CL has lexical binding). Because of that I used the macro hack lexical-let to avoid mess up the rest of code trying fix that. Lexical-let is the let of Common Lisp, whose I can easily do the things in the way I expected.