[setup.sh] remove git submodules manually on backup
git submodule is not such a thing to backup. Well, if we have
local changes not commited there is can be a problem. But just
forget about that.

Unfortunately on git version 1.9.1 we don't have `git submodule
deinit --all`. Seems to be added recently and deinit command
is very bug on old versions.