[emacs] disable setup-terminal-session call on frame creation
That way I don't have strange red colors on a full featured session
on a capable terminal, like xfce4-terminal.

But on uxterm and xterm I still have problems, as I have on sessions
without X server, which the region and helm-selection colors are a
type of gray that cannot be represented on this environment.

To fix that just call: M-x setup-terminal-session

Maybe would be better choose a better name.

The ideal config is continue to use this function on frame creation,
but only triggered on terminals without good color support.

Trying find a solution by here:
https://emacs.stackexchange.com/questions/36395/how-to-detect-the-term-color-support-inside-of-a-emacsclient-session?noredirect=1#comment57333_36395