[awesome] remove CANCER variable from autoruns.sh
I was just testing if some variable export on there will
make available on all system, a simple test to fix the problem
with SSH_AUTH_SOCK not being set. But I already fix that using
the luaposix library. Would be better a solution without external
libraries, but since I don't know (yet) will be that for while.