[gemacs] Fix properly the gemacs script raise-frames
Actually x-focus-frame is necessary because it's the function
to actually focus-frame if there is no focus on it (but it's not
minimized). raise-frames do the another part: raise properly the frame
and remove the state of minimized if is it.

So:

raise-frame: unminimize
x-focus-frame: focus the frame after unminimize

Both functions only should run on X emacs frames.