[emacs] Replace prelude-require-packages with lerax-require-packages
The intent of this it is get rid of prelude as dependency of my
configs or at least reduce it. Using directly prelude-package is no
good.

I provided an alternative implementation in my lib mano.el. It
contains now:

- lerax-require-package
- lerax-packages-installed-p
- lerax-require-packages

The main usage is lerax-require-packages, which when called
lerax-packages-installed-p and got a falsy result it will
call (package-refresh-contets) before downloading each dependency that
is not installed yet.

That way my config is more portable in new envs without errors on bootstraping.